Revert all changes - starting over from scratch
