+++ src/main/java/org/openjdk/jextract/impl/ClassSourceBuilder.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2020, 2025, Oracle and/or its affiliates. All rights reserved.
+ * Copyright (c) 2020, 2024, Oracle and/or its affiliates. All rights reserved.
  * DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS FILE HEADER.
  *
  * This code is free software; you can redistribute it and/or modify it
@@ -160,15 +160,6 @@ abstract class ClassSourceBuilder {
             """, className);
     }
 
-    final void emitPrivateConstructor() {
-        appendIndentedLines("""
-
-            private %1$s() {
-                // Should not be called directly
-            }
-            """, className);
-    }
-
     final void emitDocComment(Declaration decl) {
         emitDocComment(decl, "");
     }
@@ -209,14 +200,11 @@ abstract class ClassSourceBuilder {
         return switch (type) {
             case Primitive p -> primitiveLayoutString(p, align);
             case Declared d when Utils.isEnum(d) -> layoutString(ClangEnumType.get(d.tree()).get(), align);
-            case Declared d when Utils.isStructOrUnion(d) ->
-                    alignIfNeeded(JavaName.getFullNameOrThrow(d.tree()) + ".layout()", ClangAlignOf.getOrThrow(d.tree()) / 8, align);
-            case Delegated d when d.kind() == Delegated.Kind.POINTER ->
-                    alignIfNeeded(sb.layoutUtilsName() + "C_POINTER", 8, align);
+            case Declared d when Utils.isStructOrUnion(d) -> alignIfNeeded(JavaName.getFullNameOrThrow(d.tree()) + ".layout()", ClangAlignOf.getOrThrow(d.tree()) / 8, align);
+            case Delegated d when d.kind() == Delegated.Kind.POINTER -> alignIfNeeded(runtimeHelperName() + ".C_POINTER", 8, align);
             case Delegated d -> layoutString(d.type(), align);
-            case Function _ -> alignIfNeeded(sb.layoutUtilsName() + "C_POINTER", 8, align);
-            case Array a ->
-                    String.format("MemoryLayout.sequenceLayout(%1$d, %2$s)", a.elementCount().orElse(0L), layoutString(a.elementType(), align));
+            case Function _ -> alignIfNeeded(runtimeHelperName() + ".C_POINTER", 8, align);
+            case Array a -> String.format("MemoryLayout.sequenceLayout(%1$d, %2$s)", a.elementCount().orElse(0L), layoutString(a.elementType(), align));
             default -> throw new UnsupportedOperationException();
         };
     }
@@ -255,16 +243,16 @@ abstract class ClassSourceBuilder {
 
     private String primitiveLayoutString(Primitive primitiveType, long align) {
         return switch (primitiveType.kind()) {
-            case Bool -> sb.layoutUtilsName() + "C_BOOL";
-            case Char -> sb.layoutUtilsName() + "C_CHAR";
-            case Short -> alignIfNeeded(sb.layoutUtilsName() + "C_SHORT", 2, align);
-            case Int -> alignIfNeeded(sb.layoutUtilsName() + "C_INT", 4, align);
-            case Long -> alignIfNeeded(sb.layoutUtilsName() + "C_LONG", TypeImpl.IS_WINDOWS ? 4 : 8, align);
-            case LongLong -> alignIfNeeded(sb.layoutUtilsName() + "C_LONG_LONG", 8, align);
-            case Float -> alignIfNeeded(sb.layoutUtilsName() + "C_FLOAT", 4, align);
-            case Double -> alignIfNeeded(sb.layoutUtilsName() + "C_DOUBLE", 8, align);
+            case Bool -> runtimeHelperName() + ".C_BOOL";
+            case Char -> runtimeHelperName() + ".C_CHAR";
+            case Short -> alignIfNeeded(runtimeHelperName() + ".C_SHORT", 2, align);
+            case Int -> alignIfNeeded(runtimeHelperName() + ".C_INT", 4, align);
+            case Long -> alignIfNeeded(runtimeHelperName() + ".C_LONG", TypeImpl.IS_WINDOWS ? 4 : 8, align);
+            case LongLong -> alignIfNeeded(runtimeHelperName() + ".C_LONG_LONG", 8, align);
+            case Float -> alignIfNeeded(runtimeHelperName() + ".C_FLOAT", 4, align);
+            case Double -> alignIfNeeded(runtimeHelperName() + ".C_DOUBLE", 8, align);
             case LongDouble -> TypeImpl.IS_WINDOWS ?
-                    alignIfNeeded(sb.layoutUtilsName() + "C_LONG_DOUBLE", 8, align) :
+                    alignIfNeeded(runtimeHelperName() + ".C_LONG_DOUBLE", 8, align) :
                     paddingLayoutString(8, 0);
             case HalfFloat, Char16, WChar -> paddingLayoutString(2, 0); // unsupported
             case Float128, Int128 -> paddingLayoutString(16, 0); // unsupported
@@ -274,7 +262,7 @@ abstract class ClassSourceBuilder {
 
     private String alignIfNeeded(String layoutPrefix, long align, long expectedAlign) {
         return align > expectedAlign ?
-                String.format("%3$salign(%1$s, %2$d)", layoutPrefix, expectedAlign, sourceFileBuilder().FFMUtilsName()) :
+                String.format("%1$s.align(%2$s, %3$d)", runtimeHelperName(), layoutPrefix, expectedAlign) :
                 layoutPrefix;
     }
