+++ src/main/java/org/openjdk/jextract/impl/SourceFileBuilder.java
@@ -49,26 +49,10 @@ final class SourceFileBuilder {
         return className;
     }
 
-    public String getPackageName(){
-        return packageName;
-    }
-
-    public String layoutUtilsName() {
-        return getPackageName().isEmpty() ? ToplevelBuilder.LAYOUT_UTILS + "." : "";
-    }
-
-    public String FFMUtilsName() {
-        return getPackageName().isEmpty() ? ToplevelBuilder.FFM_UTILS + "." : "";
-    }
-
     public static SourceFileBuilder newSourceFile(String packageName, String className) {
         SourceFileBuilder sfb = new SourceFileBuilder(packageName, className);
         sfb.emitPackagePrefix();
         sfb.emitImportSection();
-        if (!className.equals(ToplevelBuilder.FFM_UTILS) && !className.equals(ToplevelBuilder.LAYOUT_UTILS)
-                && !packageName.isEmpty()) {
-            sfb.emitStaticImportSection();
-        }
         return sfb;
     }
 
@@ -97,16 +81,9 @@ final class SourceFileBuilder {
 
             import static java.lang.foreign.ValueLayout.*;
             import static java.lang.foreign.MemoryLayout.PathElement.*;
-
             """);
     }
 
-    void emitStaticImportSection() {
-        appendLines(String.format("""
-            import static %1$s.FFMUtils.*;
-            import static %1$s.LayoutUtils.*;
-            """, packageName));
-    }
 
     // Internal generation helpers (used by other builders)
