+++ test/jtreg/generator/clinitCycles/TestGlobal.java
@@ -1,4 +1,4 @@
-/* Copyright (c) 2023, 2025, Oracle and/or its affiliates. All rights reserved.
+/* Copyright (c) 2023, 2024, Oracle and/or its affiliates. All rights reserved.
  * DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS FILE HEADER.
  *
  * This code is free software; you can redistribute it and/or modify it
@@ -40,7 +40,7 @@ public class TestGlobal {
 
     @Test
     public void testGlobal() {
-        ValueLayout layout = LayoutUtils.C_INT;
+        ValueLayout layout = clinit_global_h.C_INT;
         assertNotNull(layout);
         assertEquals(clinit_global_h.global1(), 1);
         assertEquals(clinit_global_h.global2(), 2);
