+++ test/jtreg/generator/dedup/TestDedup.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2023, 2025, Oracle and/or its affiliates. All rights reserved.
+ * Copyright (c) 2023, 2024, Oracle and/or its affiliates. All rights reserved.
  * DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS FILE HEADER.
  *
  * This code is free software; you can redistribute it and/or modify it
@@ -46,7 +46,7 @@ public class TestDedup {
     }
 
     void checkLayout(GroupLayout layout) {
-        assertEquals(layout.memberLayouts().get(0), LayoutUtils.C_INT.withName("x"));
-        assertEquals(layout.memberLayouts().get(1), LayoutUtils.C_INT.withName("y"));
+        assertEquals(layout.memberLayouts().get(0), macro_fields_h.C_INT.withName("x"));
+        assertEquals(layout.memberLayouts().get(1), macro_fields_h.C_INT.withName("y"));
     }
 }
