+++ test/jtreg/generator/nestedTypes/TestNestedTypesUnsupported.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2024, 2025, Oracle and/or its affiliates. All rights reserved.
+ * Copyright (c) 2024, Oracle and/or its affiliates. All rights reserved.
  * DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS FILE HEADER.
  *
  * This code is free software; you can redistribute it and/or modify it
@@ -42,7 +42,7 @@ import static org.testng.Assert.assertTrue;
 public class TestNestedTypesUnsupported {
 
     static MemoryLayout UNDEFINED_STRUCT = MemoryLayout.structLayout(
-            MemoryLayout.paddingLayout(LayoutUtils.C_POINTER.byteSize())
+            MemoryLayout.paddingLayout(nested_types_unsupported_h.C_POINTER.byteSize())
     );
 
     @Test
