+++ test/jtreg/generator/test8244412/LibTest8244412Test.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2020, 2025, Oracle and/or its affiliates. All rights reserved.
+ * Copyright (c) 2020, 2024, Oracle and/or its affiliates. All rights reserved.
  * DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS FILE HEADER.
  *
  * This code is free software; you can redistribute it and/or modify it
@@ -31,7 +31,6 @@ import static org.testng.Assert.assertTrue;
 
 import test.jextract.test8244412.*;
 import static test.jextract.test8244412.test8244412_h.*;
-import static test.jextract.test8244412.LayoutUtils.*;
 
 /*
  * @test
