+++ test/jtreg/generator/test8252121/Test8252121.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2020, 2025, Oracle and/or its affiliates. All rights reserved.
+ * Copyright (c) 2020, 2024, Oracle and/or its affiliates. All rights reserved.
  * DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS FILE HEADER.
  *
  * This code is free software; you can redistribute it and/or modify it
@@ -46,7 +46,7 @@ public class Test8252121 {
     public void test() {
         try (var arena = Arena.ofConfined()) {
             int[] array = { 3, 5, 89, 34, -33 };
-            MemorySegment seg = arena.allocateFrom(LayoutUtils.C_INT, array);
+            MemorySegment seg = arena.allocateFrom(C_INT, array);
             assertEquals(IntStream.of(array).sum(), sum(seg));
             assertEquals(IntStream.of(array).reduce(1, (a,b) -> a*b), mul(seg));
         }
