+++ test/jtreg/generator/test8257892/LibUnsupportedTest.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2020, 2025, Oracle and/or its affiliates. All rights reserved.
+ * Copyright (c) 2020, 2024, Oracle and/or its affiliates. All rights reserved.
  * DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS FILE HEADER.
  *
  * This code is free software; you can redistribute it and/or modify it
@@ -34,8 +34,6 @@ import test.jextract.unsupported.unsupported_h;
 
 import static org.testng.Assert.*;
 import static test.jextract.unsupported.unsupported_h.*;
-import static test.jextract.unsupported.LayoutUtils.*;
-import static test.jextract.unsupported.FFMUtils.*;
 import test.jextract.unsupported.*;
 
 /*
