+++ test/jtreg/generator/testStruct/LibStructTest.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2020, 2025, Oracle and/or its affiliates. All rights reserved.
+ * Copyright (c) 2020, 2024, Oracle and/or its affiliates. All rights reserved.
  * DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS FILE HEADER.
  *
  * This code is free software; you can redistribute it and/or modify it
@@ -31,8 +31,6 @@ import org.testng.annotations.Test;
 
 import static org.testng.Assert.assertEquals;
 import static test.jextract.struct.struct_h.*;
-import static test.jextract.struct.LayoutUtils.*;
-import static test.jextract.struct.FFMUtils.*;
 import test.jextract.struct.*;
 
 /*
